move the flags around, make the http copy buffers buffered
