+++ server/configs.go
@@ -13,7 +13,7 @@ import (
 	http_director "github.com/mwitkow/kedge/http/director"
 	http_adhoc "github.com/mwitkow/kedge/http/director/adhoc"
 	http_router "github.com/mwitkow/kedge/http/director/router"
-	"github.com/mwitkow/kedge/server/sharedflags"
+	"github.com/mwitkow/kedge/lib/sharedflags"
 )
 
 var (
